requirements.txt
Django==5.0.2
djangorestframework==3.14.0
Markdown==3.5.2
django-filter==24.1
black==24.4.2